MLProcessorService/requirements.txt
aiokafka==0.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
kafka-python==2.2.15
kagglehub==0.3.12
MarkupSafe==3.0.2
minio==7.2.15
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pycparser==2.22
pycryptodome==3.23.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
segmentation_models_pytorch==0.5.0
sympy==1.14.0
timm==1.0.16
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
